refactor: move useWallet to 'hooks' directory
